Fix VDS layer visualization
Former-commit-id: 79b4f6aced74d610b10c831d10888c0bece6398c

+++ GeoImageAnalysis/src/main/java/org/geoimage/analysis/BoatConnectedPixelMap.java
@@ -19,7 +19,7 @@ public class BoatConnectedPixelMap {
         private double boatlength = 0.0;
         private double boatheading = 0.0;
         private int id = 0;
-        private double maxValue = 0.0;
+        private int maxValue = 0;
         private boolean touchlandmask = false;
         
         
@@ -98,7 +98,7 @@ public class BoatConnectedPixelMap {
             return boatwidth;
         }
 
-        protected double getMaximumValue() {
+        protected int getMaximumValue() {
             return maxValue;
         }
